refactor: output active DB version in log by default
Information about the vulnerability database in use during a scanning
operation is currently only outputted at the debug log level. This
information - especially the database's age - is useful to know, since
it provides important context for Trivy's findings. Decrease the
required verbosity level for these messages to info so it is outputted
by default.